+++ back-end/router/authors.js
@@ -18,13 +18,6 @@ router.get("/:id", (req, res) => {
   }); 
 });
 
-// router.post("/", (req, res) => {
-//   Author.create(req.body, function (err, res) {
-//     if (err) return handleError(err);
-//   });
-//   res.send('saved!')
-// });
-
 router.delete("/:id", (req, res) => {
   Author.deleteMany({ _id: req.params.id}, () => {
   console.log('deleted')
@@ -105,26 +98,5 @@ router.post('/signup', (req, res) => {
       };
   })
 });
-
-router.get('/private', (req, res) => {
-  const authHeader = req.get("Authorization");
-  if (!authHeader) {
-      return res.status(401).json({ message: 'not authenticated' });
-  };
-  const token = authHeader.split(' ')[1];
-  let decodedToken; 
-  try {
-      decodedToken = jwt.verify(token, 'secret');
-      console.log(decodedToken)
-  } catch (err) {
-      return res.status(500).json({ message: err.message || 'could not decode the token' });
-  };
-  if (!decodedToken) {
-      res.status(401).json({ message: 'unauthorized' });
-  } else {
-      res.status(200).json({ message: 'here is your resource' });
-  };
-}
-);
 
 module.exports = router;
