+++ back-end/router/users.js
@@ -1,0 +1,27 @@
+const express = require("express");
+const jwt = require('jsonwebtoken');
+let router = express.Router();
+
+
+router.get('/private', (req, res) => {
+    const authHeader = req.get("Authorization");
+    if (!authHeader) {
+        return res.status(401).json({ message: 'not authenticated' });
+    };
+    const token = authHeader.split(' ')[1];
+    let decodedToken; 
+    try {
+        decodedToken = jwt.verify(token, 'secret');
+        console.log(decodedToken)
+    } catch (err) {
+        return res.status(500).json({ message: err.message || 'could not decode the token' });
+    };
+    if (!decodedToken) {
+        res.status(401).json({ message: 'unauthorized' });
+    } else {
+        res.status(200).json({ message: 'here is your resource' });
+    };
+}
+);
+
+module.exports = router;
