--- mongoose/router/users.js
@@ -1,86 +1,0 @@
-const express = require("express");
-const bcrypt = require('bcryptjs');
-const jwt = require('jsonwebtoken');
-const mongoose = require('mongoose');
-const UserSchema = require("../usersSchema");
-
-let router = express.Router();
-const User = mongoose.model('User', UserSchema);
-
-
-router.post('/login', (req, res, next) => {
-    User.findOne({email: req.body.email}, (err, dbUser) => {
-        if (!dbUser) {
-            return res.status(404).json({message: "user not found"});
-        } else {
-            // password hash
-            bcrypt.compare(req.body.password, dbUser.password, (err, compareRes) => {
-                if (err) { // error while comparing
-                    res.status(502).json({message: "error while checking user password"});
-                } else if (compareRes) { // password match
-                    const token = jwt.sign({ email: req.body.email }, 'secret', { expiresIn: '1h' });
-                    res.status(200).json({message: "user logged in", "token": token});
-                } else { // password doesnt match
-                    res.status(401).json({message: "invalid credentials"});
-                };
-            });
-        };
-      }); 
-}
-);
-
-router.post('/signup', (req, res, next) => {
-    // checks if email already exists
-    User.findOne({email: req.body.email}, (err, dbUser)=>{
-        if (dbUser) {
-            return res.status(409).json({message: "email already exists"});
-        } else if (req.body.email && req.body.password) {
-            // password hash
-            bcrypt.hash(req.body.password, 12, (err, passwordHash) => {
-                if (err) {
-                    return res.status(500).json({message: "couldnt hash the password"}); 
-                } else if (passwordHash) {
-                    return User.create(({
-                        email: req.body.email,
-                        name: req.body.name,
-                        password: passwordHash,
-                    }))
-                    .then(() => {
-                        res.status(200).json({message: "user created"});
-                    })
-                    .catch(err => {
-                        console.log(err);
-                        res.status(502).json({message: "error while creating the user"});
-                    });
-                };
-            });
-        } else if (!req.body.password) {
-            return res.status(400).json({message: "password not provided"});
-        } else if (!req.body.email) {
-            return res.status(400).json({message: "email not provided"});
-        };
-    })
-});
-
-router.get('/private', (req, res, next) => {
-    const authHeader = req.get("Authorization");
-    if (!authHeader) {
-        return res.status(401).json({ message: 'not authenticated' });
-    };
-    const token = authHeader.split(' ')[1];
-    let decodedToken; 
-    try {
-        decodedToken = jwt.verify(token, 'secret');
-        console.log(decodedToken)
-    } catch (err) {
-        return res.status(500).json({ message: err.message || 'could not decode the token' });
-    };
-    if (!decodedToken) {
-        res.status(401).json({ message: 'unauthorized' });
-    } else {
-        res.status(200).json({ message: 'here is your resource' });
-    };
-}
-);
-
-module.exports = router;
